refactor: Rename util package to utils
