+++ src/mg/emberframework/utils/reflection/ClassUtils.java
@@ -1,4 +1,4 @@
-package mg.emberframework.util.reflection;
+package mg.emberframework.utils.reflection;
 
 import java.io.InputStream;
 import java.io.OutputStream;
