+++ src/mg/emberframework/utils/validation/ValidatorRegistry.java
@@ -1,4 +1,4 @@
-package mg.emberframework.util.validation;
+package mg.emberframework.utils.validation;
 
 import java.lang.annotation.Annotation;
 import java.util.HashMap;
@@ -9,13 +9,13 @@ import mg.emberframework.annotation.validation.Email;
 import mg.emberframework.annotation.validation.Length;
 import mg.emberframework.annotation.validation.Numeric;
 import mg.emberframework.annotation.validation.Required;
-import mg.emberframework.util.validation.validators.DateValidator;
-import mg.emberframework.util.validation.validators.EmailValidator;
-import mg.emberframework.util.validation.validators.FieldValidator;
-import mg.emberframework.util.validation.validators.LengthValidator;
-import mg.emberframework.util.validation.validators.NeutralValidator;
-import mg.emberframework.util.validation.validators.NumericValidator;
-import mg.emberframework.util.validation.validators.RequiredValidator;
+import mg.emberframework.utils.validation.validators.DateValidator;
+import mg.emberframework.utils.validation.validators.EmailValidator;
+import mg.emberframework.utils.validation.validators.FieldValidator;
+import mg.emberframework.utils.validation.validators.LengthValidator;
+import mg.emberframework.utils.validation.validators.NeutralValidator;
+import mg.emberframework.utils.validation.validators.NumericValidator;
+import mg.emberframework.utils.validation.validators.RequiredValidator;
 
 public class ValidatorRegistry {
     private ValidatorRegistry() {
