+++ src/mg/emberframework/utils/validation/validators/EmailValidator.java
@@ -1,4 +1,4 @@
-package mg.emberframework.util.validation.validators;
+package mg.emberframework.utils.validation.validators;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
