+++ src/mg/emberframework/utils/validation/validators/NeutralValidator.java
@@ -1,17 +1,15 @@
-package mg.emberframework.util.validation.validators;
+package mg.emberframework.utils.validation.validators;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 
 import mg.emberframework.manager.exception.ModelValidationException;
 
-public class RequiredValidator implements FieldValidator{
+public class NeutralValidator implements FieldValidator{
 
     @Override
     public void validate(String value, Annotation annotation, Field field) throws ModelValidationException {
-        if (value == null || "".equals(value)) {
-            throw new ModelValidationException(field.getName() + " is required");
-        }
+        // This method is empty because this validate field without validator annotation.
     }
     
 }
