+++ src/mg/emberframework/utils/validation/validators/NumericValidator.java
@@ -1,15 +1,17 @@
-package mg.emberframework.util.validation.validators;
+package mg.emberframework.utils.validation.validators;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 
 import mg.emberframework.manager.exception.ModelValidationException;
 
-public class NeutralValidator implements FieldValidator{
+public class NumericValidator implements FieldValidator{
 
     @Override
     public void validate(String value, Annotation annotation, Field field) throws ModelValidationException {
-        // This method is empty because this validate field without validator annotation.
+        if (!(value.matches("\\d+"))) {
+            throw new ModelValidationException(field.getName() + " should be numeric");
+        }
     }
     
 }
