+++ src/mg/emberframework/utils/validation/validators/RequiredValidator.java
@@ -1,16 +1,16 @@
-package mg.emberframework.util.validation.validators;
+package mg.emberframework.utils.validation.validators;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 
 import mg.emberframework.manager.exception.ModelValidationException;
 
-public class NumericValidator implements FieldValidator{
+public class RequiredValidator implements FieldValidator{
 
     @Override
     public void validate(String value, Annotation annotation, Field field) throws ModelValidationException {
-        if (!(value.matches("\\d+"))) {
-            throw new ModelValidationException(field.getName() + " should be numeric");
+        if (value == null || "".equals(value)) {
+            throw new ModelValidationException(field.getName() + " is required");
         }
     }
